pyproject.toml
[tool.poetry]
name = "test-utf-tech"
version = "0.1.0"
description = ""
authors = ["nikita-r <gfgfgf354@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
django-model-utils = "^4.5.1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
